requirements.txt
streamlit==1.37.1
sentence_transformers==3.0.1
chromadb==0.5.11
google-generativeai==0.8.0
ipython==8.18.1
tiktoken==0.7.0
nltk==3.8.1
python-dotenv==1.0.1
langchain-experimental==0.3.2
langchain-community==0.3.3
langchain-google-genai==2.0.1
networkx
pyvis==0.3.2
streamlit chromadb bcrypt